formatted code for happy number
